Changed the user controller to Reservable controller

+++ server/src/main/java/nl/tudelft/oopp/demo/repositories/FoodCourtRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface FoodCourtRepository extends JpaRepository<Foodcourt, Integer>{
+public interface FoodCourtRepository extends JpaRepository<Foodcourt, Long>{
 }
