+++ server/src/main/java/nl/tudelft/oopp/demo/repositories/FoodReservationsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FoodReservationsRepository extends JpaRepository<FoodReservations, Integer>{
+public interface FoodReservationsRepository extends JpaRepository<FoodReservations, Long>{
 }
